Changed structure and added 2 functions
Changed structure of the notebook to comply with the CRISP-DM process. Included section titles on the notebook, business understanding incl. the questions asked, sorted out and made sections that contain the data cleaning and exploration process. Added bolded titles of these sections. 

Added 2 functions that reduce the repetition when printing and plotting data.